pyproject.toml
[project]
name = "prediction-market-backtester"
version = "0.1.0"
description = "Quant-style backtesting engine for prediction markets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "polars>=1.20.0",
  "pydantic>=2.10.0",
  "pyarrow>=18.0.0",
  "numpy>=2.2.0",
  "matplotlib>=3.10.0",
  "pyyaml>=6.0",
]

[project.scripts]
pm-bt = "pm_bt.cli:main"

[dependency-groups]
dev = [
  "ruff>=0.15.0",
  "pytest>=9.0.2",
  "pre-commit>=4.5.1",
  "basedpyright>=1.37.4",
]
index = [
  "httpx>=0.28.1",
  "tenacity>=9.0.0",
  "duckdb>=1.4.1",
  "pandas>=2.3.3",
  "tqdm>=4.67.1",
  "web3>=7.14.0",
  "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pm_bt"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["vendor"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.basedpyright]
pythonVersion = "3.11"
typeCheckingMode = "recommended"
include = ["src", "tests"]
exclude = ["vendor"]
venvPath = "."
venv = ".venv"
extraPaths = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
